requirements_carles.txt
absl-py==2.1.0
appdirs==1.4.4
# asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
astunparse==1.6.3
# backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
beautifulsoup4==4.12.3
bigquery==0.0.41
bs4==0.0.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
# colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
# comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.1.1
cryptography==43.0.0
cssselect==1.2.0
cycler==0.12.1
dacktool==0.0.7
db-dtypes==1.2.0
dbstream==0.1.25
# debugpy @ file:///C:/b/abs_c0y1fjipt2/croot/debugpy_1690906864587/work
# decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
# executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
fake-useragent==1.5.1
feedparser==6.0.11
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozendict==2.4.4
fsspec==2024.6.1
gast==0.4.0
google-api-core==2.19.1
google-api-python-client==1.7.11
google-auth==2.32.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.25.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.7.2
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
googleauthentication==0.0.17
grpc-google-iam-v1==0.12.7
grpcio==1.65.2
grpcio-status==1.62.2
h5py==3.11.0
html5lib==1.1
httplib2==0.22.0
huggingface-hub==0.24.5
idna==3.7
# importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1721856510709/work
importlib_resources==6.4.0
# ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
# ipython @ file:///D:/bld/ipython_1680185618122/work
# jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.4
joblib==1.4.2
# jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1716472197302/work
# jupyter_core @ file:///D:/bld/jupyter_core_1710257325428/work
keras==2.13.1
kiwisolver==1.4.5
libclang==18.1.1
libcst==1.1.0
lxml==5.2.2
lxml_html_clean==0.2.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
# matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mpmath==1.3.0
multitasking==0.0.11
mypy-extensions==1.0.0
# nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.1
nltk==3.8.1
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
opt-einsum==3.3.0
# packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==2.0.3
pandas-gbq==0.23.1
parse==1.20.2
# parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
peewee==3.17.6
# pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.4.0
# platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
# prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
proto-plus==1.24.0
protobuf==4.25.4
# psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
# pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
py4j==0.10.9.7
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydata-google-auth==1.8.2
PyDrive==1.3.1
pyee==11.1.0
# Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyparsing==3.1.2
pyppeteer==2.0.0
pyquery==2.0.0
pyspark==3.5.2
# python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytz==2024.1
# pywin32==305.1
PyYAML==6.0.1
# pyzmq @ file:///C:/b/abs_89aq69t0up/croot/pyzmq_1705605705281/work
regex==2024.7.24
requests==2.32.3
requests-html==0.10.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.4
scikit-learn==1.3.2
scipy==1.10.1
sgmllib3k==1.0.0
# six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
soupsieve==2.5
# stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
stocksent==0.4.5
sympy==1.13.2
# TA-Lib @ file:///D:/bld/ta-lib_1719846510552/work
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
torchvision==0.19.0
# tornado @ file:///D:/bld/tornado_1656937958156/work
tqdm==4.66.4
# traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
transformers==4.44.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==3.0.1
urllib3==1.26.19
w3lib==2.2.1
# wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
webencodings==0.5.1
websockets==10.4
Werkzeug==3.0.3
wordcloud==1.9.3
wrapt==1.16.0
yahoo-fin==0.8.9.1
yfinance==0.2.41
# zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1718013267051/work
